requirements.txt
pytesseract==0.3.10
pillow==10.2.0
numpy==1.26.4
ollama==0.1.5
python-dotenv==1.0.1 